Add -print-hitmiss option
This option is added for knowing if an access hit or miss the cache.
The hit (H) and miss (M) per reference is printed to the standard
output.

This option is not perfectly implemented. If an access cross a
cacheline, multiple H or M will be printed. Also, if there are
multiple cache layers, e.g., L1D->L2U, multiple H or M may be
printed as well.